requirements.txt
Flask==2.1.0
Flask-CORS==3.0.10
gunicorn==20.1.0
python-telegram-bot==13.7
torch==1.8.1
numpy==1.22.3
nltk==3.7.1
requests==2.27.1
python-dotenv
redis